Move renderer process files into separate directory. Change directories structure for services/events
