+++ kiosk/src/admin_panel.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>ScreenDriver</title>
-    <script src="js/admin_panel.js"></script>
+    <script src="js/renderer/admin_panel.js"></script>
     <link rel="stylesheet" href="../node_modules/bootstrap/dist/css/bootstrap.min.css">
     <link rel="stylesheet" href="../node_modules/bootstrap/dist/css/bootstrap-theme.min.css">
     <link rel="stylesheet" href="../node_modules/bootstrap-select/dist/css/bootstrap-select.min.css">
