+++ kiosk/src/js/services/events/internal-events-listeners/hotkeys_events_handler.js
@@ -1,8 +1,8 @@
 const electron = require('electron');
 const app = electron.app;
-const CronJobsManager = require('../../helpers/settings_load_job_manager');
-const {scheduledTaskManager} = require('./../../scheduled-task-manager');
-const WindowsHelper = require('./../../helpers/windows_helper');
+const CronJobsManager = require('../../../helpers/settings_load_job_manager');
+const {scheduledTaskManager} = require('../../../scheduled-task-manager');
+const WindowsHelper = require('../../../helpers/windows_helper');
 
 
 /**
