+++ kiosk/src/js/services/events/internal-events-listeners/renderer_events_handler.js
@@ -1,9 +1,9 @@
 const {ipcMain} = require('electron');
 
-const CurrentScreenSettingsManager = require('../../current_screen_settings_manager');
-const Logger = require('../../logger/logger');
-const UserInteractionsManager = require('../../user-interactions-manager');
-const WindowsHelper = require('./../../helpers/windows_helper');
+const CurrentScreenSettingsManager = require('../../../current_screen_settings_manager');
+const Logger = require('../../../logger/logger');
+const UserInteractionsManager = require('../../../user-interactions-manager');
+const WindowsHelper = require('../../../helpers/windows_helper');
 
 /**
  * Should be used after application's 'ready' event was triggered
