+++ kiosk/src/js/services/services_initialiser.js
@@ -14,9 +14,9 @@ class ServicesInitialiser {
 
     //should be called AFTER application was initialized and ready event was called
     static initBehaviourServices() {
-        require('./external-event-listeners/api_events_handler_service').init();
-        require('./events/renderer_events_handler');
-        require('./events/hotkeys_events_handler');
+        require('./events/external-event-listeners/api_events_handler_service').init();
+        require('./events/internal-events-listeners/renderer_events_handler');
+        require('./events/internal-events-listeners/hotkeys_events_handler');
     }
 }
